Tue, Mar 5, 2024, 12:03 AM +01:00

+++ finance/app.py
@@ -55,7 +55,7 @@ def index():
     userEntry = db.execute("SELECT cash FROM users WHERE id = ?", session.get("user_id"))
     if len(userEntry) != 1:
         return apology("Internal Server Error", 500)
-    
+
     return render_template("index.html", portfolio=portfolio, sum=usd(sum), cash=usd(userEntry[0]['cash']))
 
 
@@ -207,7 +207,7 @@ def sell():
         if len(user_cash) != 1:
             return apology("Something went wrong", 500)
         amount_possesion = db.execute("SELECT SUM(amount) as amount FROM transactions WHERE user_id =? AND symbol = ?", session.get("user_id"), selected)
-        new_saldo = int(user_cash[0]['cash'])+int(amount_possesion[0]['amount']) * quote['price']
+        new_saldo = int(user_cash[0]['cash']) + float(amount_possesion[0]['amount']) * quote['price']
         db.execute("UPDATE users SET cash = ? WHERE id = ?",new_saldo,session.get("user_id"))
         db.execute("INSERT INTO transactions (user_id, symbol, amount, quote, type)VALUES (?, ?, ?, ?, 'sell')", session.get("user_id"), selected, amount_possesion[0]['amount'], quote['price'])
         return redirect("/")
